+++ ServiceLayer/Services/PanelServices/IPanelServices.cs
@@ -10,7 +10,7 @@ namespace ServiceLayer.Services.PanelServices
     {
         void ValidateModel(IPanelModel panelModel);
         IPanelModel AddPanelModel(int panelWd,
-                                  double panelWdWithDecimal,
+                                  decimal panelWdWithDecimal,
                                   int panelHt,
                                   Control panelParent,
                                   UserControl panelFrameGroup,
